Commit changes

+++ src/main/java/com/example/recs/domain/dto/RatingDto.java
@@ -11,6 +11,6 @@ import lombok.NoArgsConstructor;
 @Builder
 public class RatingDto {
     private long userId;
-    private long movieId;
+    private Integer movieId;
     private double rating;
 }
